+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Version 0.9.2 (à venir) :
  - Réorganisation interne des fichiers.
  - Passage au framework .NET 4.7
+ - Ajout d'une option pour obtenir la liste des fichiers et pour sélectionner un fichier précis dans la liste.
 
 ### Version 0.9.1 (12 Mars 2017) :
  - Corrections de quelques bugs (parce qu'il y en a toujours).